Leverage causality in consistent backend
* Its Reads are now also using IVersion comparables, instead
of transmitting entire contents of everything read.
* Every server, initiator or receiver, will now make sure to
create dependency links between the transaction and all its
Reads. Applying a transaction thus no longer needs to apply
Reads. The dependency links are established by the Check
method.
* Minor: Fix namespaces of Utils and TransactionalTickCount.
* Small optimization for direct mail. Avoid creating package
if it won't be needed.
* If received direct mail has unmet dependencies, we will now
start a gossip exchange with sender. Configurable behavior.
This is important for transactions that get rejected, which
creates a dependency on the version that caused the rejection,
and obviously that dependency will be unmet on the initiator.
* Stricter check for when an old waitingStore may be used.
* Some optimizations for reducing needless dependencies.