First test and fix for causal consistency
The test runs 10k transactions that read from one field
and save its key and version into another field. And every
time it reads anything, it checks that the read key/version
match what he can see under that other key. Additionally,
a Changed event checks them all any time any change
happens.

The test quickly identified that the old algorithm was
broken. The "future dependencies" business was very
wrong. The assumption that any dependency that our
future dependency has would also be our dependency,
was wrong. A future version can get more, which did not
exist when we read its old version. And so the trans that
depends on such a "future" dependency can get applied,
while that "future" key it depends on does not.

The algorithm is now much cleaner, and no longer
depends on transitively closing dependencies. Only when
doing writes, we have to pull all old dependencies of
the field we're writing into, because the old version will
be gone once we commit, and thus that dependency info
would be gone with it. And when just reading a field, we
only need a dependency on that field and the version we
read.