Improve gossip message creation
Split the message creating methods into two Shielded
transactions, so that the package preparing, which
iterates over the _freshIndex and is exetremely prone to
conflicts, happens as much as possible in a read-only
Shielded transaction.

End messages still have to change the state in the same
transaction, because they must be retired if something
new happens. Delaying a state change in that case could
lead to the new changes not being sent to the other
server. So this change really affects situations where our
response is a GossipReply.